fix(ui/float): clear EndOfBuffer background
When a background color is set in the EndOfBuffer highlight group, the
bottom part of floating windows which content does not fill the entire
window height appears truncated.

This can easily be observed by opening a lazy.nvim window (plugin
manager for Neovim), or mason.nvim window (tools installer for Neovim)
with a filter applied.

Fixes #123