fix(lua/lsp): built-in variables highlighted as Fg
LSP Semantic Tokens aren't expressive enough to apply special modifiers
to built-in variables (such as 'self' in Lua) and identifies them as any
other variable.

Since this causes a degradation of the syntax highlighting in comparison
with Tree-sitter (which has '@variable.builtin'), we fall back to
Tree-sitter by clearing the '@lsp.type.variable' highlight group in Lua
files.

Fixes #155